requirements.txt
streamlit==1.46.0
numpy==1.26.4
pandas==2.2.1
plotly==6.1.2
requests==2.32.3
matplotlib==3.8.3
beautifulsoup4==4.13.4
textblob==0.19.0
transformers==4.47.0
tensorflow==2.19.0
scikit-learn==1.7.0
backtrader==1.9.78.123
python-dateutil==2.8.2
lxml==5.4.0
python-dotenv~=1.1.1
tf-keras==2.19.0
